+++ src/Common/SchemaExtractor.php
@@ -1,25 +1,23 @@
 <?php
 
-
 namespace TgScraper\Common;
 
-
 use Composer\InstalledVersions;
+use Exception;
+use GuzzleHttp\Client;
+use GuzzleHttp\Exception\GuzzleException;
 use InvalidArgumentException;
 use JetBrains\PhpStorm\ArrayShape;
 use OutOfBoundsException;
-use PHPHtmlParser\Dom;
-use PHPHtmlParser\Exceptions\ChildNotFoundException;
-use PHPHtmlParser\Exceptions\CircularException;
-use PHPHtmlParser\Exceptions\ContentLengthException;
-use PHPHtmlParser\Exceptions\LogicalException;
-use PHPHtmlParser\Exceptions\NotLoadedException;
-use PHPHtmlParser\Exceptions\ParentNotFoundException;
-use PHPHtmlParser\Exceptions\StrictException;
-use Psr\Http\Client\ClientExceptionInterface;
 use Psr\Log\LoggerInterface;
+use RuntimeException;
+use TgScraper\Parsers\Field;
+use TgScraper\Parsers\ObjectDescription;
 use TgScraper\Constants\Versions;
-use Throwable;
+use voku\helper\HtmlDomParser;
+use voku\helper\SimpleHtmlDomInterface;
+use voku\helper\SimpleHtmlDomNode;
+use voku\helper\SimpleHtmlDomNodeInterface;
 
 /**
  * Class SchemaExtractor
@@ -27,15 +25,6 @@ use Throwable;
  */
 class SchemaExtractor
 {
-
-    /**
-     * Additional methods with boolean return value.
-     */
-    private const BOOL_RETURNS = [
-        'answerShippingQuery',
-        'answerPreCheckoutQuery'
-    ];
-
     /**
      * @var string
      */
@@ -44,11 +33,9 @@ class SchemaExtractor
     /**
      * SchemaExtractor constructor.
      * @param LoggerInterface $logger
-     * @param Dom $dom
-     * @throws ChildNotFoundException
-     * @throws NotLoadedException
+     * @param HtmlDomParser $dom
      */
-    public function __construct(private LoggerInterface $logger, private Dom $dom)
+    public function __construct(private LoggerInterface $logger, private HtmlDomParser $dom)
     {
         $this->version = $this->parseVersion();
         $this->logger->info('Bot API version: ' . $this->version);
@@ -60,13 +47,17 @@ class SchemaExtractor
      * @param string $version
      * @return SchemaExtractor
      * @throws OutOfBoundsException
-     * @throws Throwable
+     * @throws Exception
+     * @throws GuzzleException
      */
     public static function fromVersion(LoggerInterface $logger, string $version = Versions::LATEST): SchemaExtractor
     {
         if (InstalledVersions::isInstalled('sysbot/tgscraper-cache') and class_exists('\TgScraper\Cache\CacheLoader')) {
             $logger->info('Cache package detected, searching for a cached version.');
             try {
+                /** @noinspection PhpFullyQualifiedNameUsageInspection */
+                /** @noinspection PhpUndefinedNamespaceInspection */
+                /** @psalm-suppress UndefinedClass */
                 $path = \TgScraper\Cache\CacheLoader::getCachedVersion($version);
                 $logger->info('Cached version found.');
                 return self::fromFile($logger, $path);
@@ -83,20 +74,20 @@ class SchemaExtractor
      * @param LoggerInterface $logger
      * @param string $path
      * @return SchemaExtractor
-     * @throws Throwable
+     * @throws InvalidArgumentException
+     * @throws RuntimeException
      */
     public static function fromFile(LoggerInterface $logger, string $path): SchemaExtractor
     {
-        $dom = new Dom;
-        if (!file_exists($path)) {
+        if (!file_exists($path) or is_dir($path)) {
             throw new InvalidArgumentException('File not found');
         }
         $path = realpath($path);
         try {
             $logger->info(sprintf('Loading data from file "%s".', $path));
-            $dom->loadFromFile($path);
+            $dom = HtmlDomParser::file_get_html($path);
             $logger->info('Data loaded.');
-        } catch (Throwable $e) {
+        } catch (RuntimeException $e) {
             $logger->critical(sprintf('Unable to load data from "%s": %s', $path, $e->getMessage()));
             throw $e;
         }
@@ -107,20 +98,15 @@ class SchemaExtractor
      * @param LoggerInterface $logger
      * @param string $url
      * @return SchemaExtractor
-     * @throws ChildNotFoundException
-     * @throws CircularException
-     * @throws ClientExceptionInterface
-     * @throws ContentLengthException
-     * @throws LogicalException
-     * @throws StrictException
-     * @throws NotLoadedException
+     * @throws GuzzleException
      */
     public static function fromUrl(LoggerInterface $logger, string $url): SchemaExtractor
     {
-        $dom = new Dom;
+        $client = new Client();
         try {
-            $dom->loadFromURL($url);
-        } catch (Throwable $e) {
+            $html = $client->get($url)->getBody();
+            $dom = HtmlDomParser::str_get_html((string)$html);
+        } catch (GuzzleException $e) {
             $logger->critical(sprintf('Unable to load data from URL "%s": %s', $url, $e->getMessage()));
             throw $e;
         }
@@ -129,57 +115,56 @@ class SchemaExtractor
     }
 
     /**
-     * @throws ParentNotFoundException
-     * @throws ChildNotFoundException
+     * @param SimpleHtmlDomInterface $node
+     * @return array{description: string, table: ?SimpleHtmlDomNodeInterface, extended_by: string[]}
      */
-    #[ArrayShape(['description' => "string", 'table' => "mixed", 'extended_by' => "array"])]
-    private static function parseNode(Dom\Node\AbstractNode $node): ?array
+    private static function parseNode(SimpleHtmlDomInterface $node): array
     {
         $description = '';
         $table = null;
         $extendedBy = [];
         $tag = '';
         $sibling = $node;
-        while (!str_starts_with($tag, 'h')) {
-            $sibling = $sibling->nextSibling();
-            $tag = $sibling?->tag?->name();
-            if (empty($node->text()) or empty($tag) or $tag == 'text') {
+        while (!str_starts_with($tag ?? '', 'h')) {
+            $sibling = $sibling?->nextSibling();
+            $tag = $sibling?->tag;
+            if (empty($node->text()) or empty($tag) or $tag == 'text' or empty($sibling)) {
                 continue;
-            } elseif ($tag == 'p') {
-                $description .= PHP_EOL . $sibling->innerHtml();
-            } elseif ($tag == 'ul') {
-                $items = $sibling->find('li');
-                /* @var Dom\Node\AbstractNode $item */
-                foreach ($items as $item) {
-                    $extendedBy[] = $item->innerText;
-                }
-                break;
-            } elseif ($tag == 'table') {
-                $table = $sibling->find('tbody')->find('tr');
-                break;
+            }
+            switch ($tag) {
+                case 'p':
+                    $description .= PHP_EOL . $sibling->innerHtml();
+                    break;
+                case 'ul':
+                    $items = $sibling->findMulti('li');
+                    foreach ($items as $item) {
+                        $extendedBy[] = $item->text();
+                    }
+                    break 2;
+                case 'table':
+                    /** @var SimpleHtmlDomNodeInterface $table */
+                    $table = $sibling->findOne('tbody')->findMulti('tr');
+                    break 2;
             }
         }
         return ['description' => $description, 'table' => $table, 'extended_by' => $extendedBy];
     }
 
     /**
-     * @throws ChildNotFoundException
-     * @throws NotLoadedException
+     * @return string
      */
     private function parseVersion(): string
     {
-        /** @var Dom\Node\AbstractNode $element */
-        $element = $this->dom->find('h3')[0];
+        $element = $this->dom->findOne('h3');
         $tag = '';
-        while ($tag != 'p') {
-            try {
-                $element = $element->nextSibling();
-            } catch (ChildNotFoundException | ParentNotFoundException) {
-                continue;
-            }
-            $tag = $element->tag->name();
+        while ($tag != 'p' and !empty($element)) {
+            $element = $element->nextSibling();
+            $tag = $element?->tag;
         }
-        $versionNumbers = explode('.', str_replace('Bot API ', '', $element->innerText));
+        if (empty($element)) {
+            return '1.0.0';
+        }
+        $versionNumbers = explode('.', str_replace('Bot API ', '', $element->text()));
         return sprintf(
             '%s.%s.%s',
             $versionNumbers[0] ?? '1',
@@ -189,22 +174,26 @@ class SchemaExtractor
     }
 
     /**
-     * @return array
-     * @throws Throwable
+     * @return string
      */
-    #[ArrayShape(['version' => "string", 'methods' => "array", 'types' => "array"])]
+    public function getVersion(): string
+    {
+        return $this->version;
+    }
+
+    /**
+     * @return array{version: string, methods: array, types: array}
+     * @throws Exception
+     */
     public function extract(): array
     {
-        try {
-            $elements = $this->dom->find('h4');
-        } catch (Throwable $e) {
-            $this->logger->critical(sprintf('Unable to parse data: %s', $e->getMessage()));
-            throw $e;
+        $elements = $this->dom->findMultiOrFalse('h4');
+        if (false === $elements) {
+            throw new Exception('Unable to fetch required DOM nodes');
         }
-        $data = ['version' => $this->version];
-        /* @var Dom\Node\AbstractNode $element */
+        $data = ['version' => $this->version, 'methods' => [], 'types' => []];
         foreach ($elements as $element) {
-            if (!str_contains($name = $element->text, ' ')) {
+            if (!str_contains($name = $element->text(), ' ')) {
                 $isMethod = lcfirst($name) == $name;
                 $path = $isMethod ? 'methods' : 'types';
                 ['description' => $description, 'table' => $table, 'extended_by' => $extendedBy] = self::parseNode(
@@ -225,21 +214,15 @@ class SchemaExtractor
     /**
      * @param string $name
      * @param string $description
-     * @param Dom\Node\Collection|null $unparsedFields
+     * @param SimpleHtmlDomNodeInterface|null $unparsedFields
      * @param array $extendedBy
      * @param bool $isMethod
      * @return array
-     * @throws ChildNotFoundException
-     * @throws CircularException
-     * @throws ContentLengthException
-     * @throws LogicalException
-     * @throws NotLoadedException
-     * @throws StrictException
      */
     private static function generateElement(
         string $name,
         string $description,
-        ?Dom\Node\Collection $unparsedFields,
+        ?SimpleHtmlDomNodeInterface $unparsedFields,
         array $extendedBy,
         bool $isMethod
     ): array {
@@ -250,10 +233,8 @@ class SchemaExtractor
             'fields' => $fields
         ];
         if ($isMethod) {
-            $returnTypes = self::parseReturnTypes($description);
-            if (empty($returnTypes) and in_array($name, self::BOOL_RETURNS)) {
-                $returnTypes[] = 'bool';
-            }
+            $description = new ObjectDescription($description);
+            $returnTypes = $description->getTypes();
             $result['return_types'] = $returnTypes;
             return $result;
         }
@@ -262,126 +243,33 @@ class SchemaExtractor
     }
 
     /**
-     * @param Dom\Node\Collection|null $fields
+     * @param SimpleHtmlDomNodeInterface|null $fields
      * @param bool $isMethod
      * @return array
      */
-    private static function parseFields(?Dom\Node\Collection $fields, bool $isMethod): array
+    private static function parseFields(?SimpleHtmlDomNodeInterface $fields, bool $isMethod): array
     {
         $parsedFields = [];
-        $fields = $fields ?? [];
+        $fields ??= [];
+        /** @var SimpleHtmlDomInterface $field */
         foreach ($fields as $field) {
-            /* @var Dom\Node\AbstractNode $fieldData */
-            $fieldData = $field->find('td');
-            $name = $fieldData[0]->text;
+            /** @var SimpleHtmlDomNode $fieldData */
+            $fieldData = $field->findMulti('td');
+            $name = $fieldData[0]->text();
             if (empty($name)) {
                 continue;
             }
-            $parsedData = [
-                'name' => $name,
-                'type' => strip_tags($fieldData[1]->innerHtml)
-            ];
-            $parsedData['types'] = self::parseFieldTypes($parsedData['type']);
-            unset($parsedData['type']);
+            $types = $fieldData[1]->text();
             if ($isMethod) {
-                $parsedData['optional'] = $fieldData[2]->text != 'Yes';
-                $parsedData['description'] = htmlspecialchars_decode(
-                    strip_tags($fieldData[3]->innerHtml ?? $fieldData[3]->text ?? ''),
-                    ENT_QUOTES
-                );
+                $optional = $fieldData[2]->text() != 'Yes';
+                $description = $fieldData[3]->innerHtml();
             } else {
-                $description = htmlspecialchars_decode(strip_tags($fieldData[2]->innerHtml), ENT_QUOTES);
-                $parsedData['optional'] = str_starts_with($description, 'Optional.');
-                $parsedData['description'] = $description;
+                $description = $fieldData[2]->innerHtml();
+                $optional = str_starts_with($fieldData[2]->text(), 'Optional.');
             }
-            $parsedFields[] = $parsedData;
+            $field = new Field($name, $types, $optional, $description);
+            $parsedFields[] = $field->toArray();
         }
         return $parsedFields;
     }
-
-    /**
-     * @param string $rawType
-     * @return array
-     */
-    private static function parseFieldTypes(string $rawType): array
-    {
-        $types = [];
-        foreach (explode(' or ', $rawType) as $rawOrType) {
-            if (stripos($rawOrType, 'array') === 0) {
-                $types[] = str_replace(' and', ',', $rawOrType);
-                continue;
-            }
-            foreach (explode(' and ', $rawOrType) as $unparsedType) {
-                $types[] = $unparsedType;
-            }
-        }
-        $parsedTypes = [];
-        foreach ($types as $type) {
-            $type = trim(str_replace(['number', 'of'], '', $type));
-            $multiplesCount = substr_count(strtolower($type), 'array');
-            $parsedType = trim(
-                str_replace(
-                    ['Array', 'Integer', 'String', 'Boolean', 'Float', 'True'],
-                    ['', 'int', 'string', 'bool', 'float', 'bool'],
-                    $type
-                )
-            );
-            for ($i = 0; $i < $multiplesCount; $i++) {
-                $parsedType = sprintf('Array<%s>', $parsedType);
-            }
-            $parsedTypes[] = $parsedType;
-        }
-        return $parsedTypes;
-    }
-
-    /**
-     * @param string $description
-     * @return array
-     * @throws ChildNotFoundException
-     * @throws CircularException
-     * @throws NotLoadedException
-     * @throws StrictException
-     * @throws ContentLengthException
-     * @throws LogicalException
-     * @noinspection PhpUndefinedFieldInspection
-     */
-    private static function parseReturnTypes(string $description): array
-    {
-        $returnTypes = [];
-        $phrases = explode('.', $description);
-        $phrases = array_filter(
-            $phrases,
-            function ($phrase) {
-                return (false !== stripos($phrase, 'returns') or false !== stripos($phrase, 'is returned'));
-            }
-        );
-        foreach ($phrases as $phrase) {
-            $dom = new Dom;
-            $dom->loadStr($phrase);
-            $a = $dom->find('a');
-            $em = $dom->find('em');
-            foreach ($a as $element) {
-                if ($element->text == 'Messages') {
-                    $returnTypes[] = 'Array<Message>';
-                    continue;
-                }
-
-                $multiplesCount = substr_count(strtolower($phrase), 'array');
-                $returnType = $element->text;
-                for ($i = 0; $i < $multiplesCount; $i++) {
-                    $returnType = sprintf('Array<%s>', $returnType);
-                }
-                $returnTypes[] = $returnType;
-            }
-            foreach ($em as $element) {
-                if (in_array($element->text, ['False', 'force', 'Array'])) {
-                    continue;
-                }
-                $type = str_replace(['True', 'Int', 'String'], ['bool', 'int', 'string'], $element->text);
-                $returnTypes[] = $type;
-            }
-        }
-        return $returnTypes;
-    }
-
 }