+++ src/Constants/Versions.php
@@ -1,12 +1,9 @@
 <?php
 
-
 namespace TgScraper\Constants;
-
 
 class Versions
 {
-
     public const V100 = '1.0.0';
     public const V110 = '1.1.0';
     public const V140 = '1.4.0';
@@ -43,8 +40,10 @@ class Versions
     public const V530 = '5.3.0';
     public const V540 = '5.4.0';
     public const V550 = '5.5.0';
+    public const V560 = '5.6.0';
+    public const V570 = '5.7.0';
     public const LATEST = 'latest';
-    public const STABLE = self::V550;
+    public const STABLE = self::V570;
 
     public const URLS = [
         self::V100 => 'https://web.archive.org/web/20150714025308id_/https://core.telegram.org/bots/api/',
@@ -83,6 +82,8 @@ class Versions
         self::V530 => 'https://web.archive.org/web/20210626142851id_/https://core.telegram.org/bots/api',
         self::V540 => 'https://web.archive.org/web/20211105152638id_/https://core.telegram.org/bots/api',
         self::V550 => 'https://web.archive.org/web/20211211002657id_/https://core.telegram.org/bots/api',
+        self::V560 => 'https://web.archive.org/web/20220105131529id_/https://core.telegram.org/bots/api',
+        self::V570 => 'https://web.archive.org/web/20220206103922id_/https://core.telegram.org/bots/api',
         self::LATEST => 'https://core.telegram.org/bots/api'
     ];
 
@@ -101,5 +102,4 @@ class Versions
         $version = self::getVersionFromText($text);
         return self::URLS[$version] ?? self::URLS[self::LATEST];
     }
-
 }